+++ docs/conf.py
@@ -79,7 +79,7 @@ release = version
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-exclude_patterns = ['_build']
+exclude_patterns = ['_build', '_themes/README.rst']
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.
@@ -118,6 +118,7 @@ rst_prolog = """
 .. |E| replace:: :math:`E`
 .. |F| replace:: :math:`F`
 .. |AB| replace:: :math:`AB`
+.. |AC| replace:: :math:`AC`
 .. |CD| replace:: :math:`CD`
 .. |DE| replace:: :math:`DE`
 .. |FG| replace:: :math:`FG`
@@ -129,8 +130,10 @@ rst_prolog = """
 .. |n2| replace:: :math:`n_{2}`
 .. |n0, n1| replace:: :math:`n_{0}n_{1}`
 .. |n0, n2| replace:: :math:`n_{0}n_{2}`
+.. |n0, n4| replace:: :math:`n_{0}n_{4}`
 .. |n1, n2| replace:: :math:`n_{1}n_{2}`
 .. |n0, n1, n2| replace:: :math:`n_{0}n_{1}n_{2}`
+.. |n0, n1, n4| replace:: :math:`n_{0}n_{1}n_{4}`
 .. |(AB / DE) x ([] / C)| replace:: :math:`\\frac{AB}{DE} \\times \\frac{\left[\\;\\right]}{C}`
 .. |(A / CD) x ([] / E)| replace:: :math:`\\frac{A}{CD} \\times \\frac{\left[\\;\\right]}{E}`
 .. |([] / C) x (A / D)| replace:: :math:`\\frac{\\left[\\;\\right]}{C} \\times \\frac{A}{D}`
@@ -153,6 +156,8 @@ rst_prolog = """
 .. |j| replace:: :math:`j`
 .. |i,jth| replace:: :math:`i,j^{\\textrm{th}}`
 .. |k| replace:: :math:`k`
+.. |2-D| replace:: :math:`2`-:math:`D`
+.. |N-D| replace:: :math:`N`-:math:`D`
 .. |past| replace:: :const:`DIRECTIONS[PAST]`
 .. |future| replace:: :const:`DIRECTIONS[FUTURE]`
 .. |PRECISION| replace:: :const:`constants.PRECISION`
@@ -184,6 +189,7 @@ rst_prolog = """
 .. |Concept| replace:: :class:`~pyphi.models.concept.Concept`
 .. |Constellation| replace:: :class:`~pyphi.models.concept.Constellation`
 .. |Cut| replace:: :class:`~pyphi.models.cuts.Cut`
+.. |Bipartition| replace:: :class:`~pyphi.models.cuts.Bipartition`
 .. |Mip| replace:: :class:`~pyphi.models.concept.Mip`
 .. |Mice| replace:: :class:`~pyphi.models.concept.Mice`
 .. |Node| replace:: :class:`~pyphi.node.Node`
@@ -206,7 +212,7 @@ napoleon_use_admonition_for_notes = False
 napoleon_use_admonition_for_references = False
 napoleon_use_ivar = False
 napoleon_use_param = True
-napoleon_use_rtype = True
+napoleon_use_rtype = False
 
 # -- Options for HTML output ----------------------------------------------
 
@@ -222,7 +228,7 @@ if not on_rtd:  # only import and set the theme if we're building docs locally
 
 # (Optional) Logo. Should be small enough to fit the navbar (ideally 24x24).
 # Path should be relative to the ``_static`` files directory.
-html_logo = "phi.png"
+html_logo = "_static/phi.png"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the