+++ pyphi/network.py
@@ -15,136 +15,137 @@ from . import cache, convert, utils, validate
 from .constants import DIRECTIONS, FUTURE, PAST
 
 
-# TODO!!! raise error if user tries to change TPM or CM, double-check and
-# document that states can be changed
 class Network:
     """A network of nodes.
 
     Represents the network we're analyzing and holds auxilary data about it.
 
+    Args:
+        tpm (np.ndarray): The transition probability matrix of the network.
+
+            The TPM can be provided in either state-by-node (either |2-D| or
+            |N-D|) or state-by-state form. In either form, row indices must
+            follow the **LOLI** convention (see discussion in the |examples|
+            module.) In state-by-state form column indices must also follow
+            **LOLI** convention.
+
+            If given in state-by-node form, the TPM can be either
+            2-dimensional, so that ``tpm[i]`` gives the probabilities of each
+            node being on if the past state is encoded by |i| according to
+            **LOLI**, or in |N-D| form, so that ``tpm[(0, 0, 1)]`` gives the
+            probabilities of each node being on if the past state is
+            |N_0 = 0, N_1 = 0, N_2 = 1|.
+
+            The shape of the 2-dimensional form of a state-by-node TPM must be
+            ``(S, N)``, and the shape of the |N-D| form of the TPM must be
+            ``[2] * N + [N]``, where ``S`` is the number of states and |N| is
+            the number of nodes in the network.
+
+    Keyword Args:
+        connectivity_matrix (np.ndarray): A square binary adjacency matrix
+            indicating the connections between nodes in the network.
+            ``connectivity_matrix[i][j] == 1`` means that node |i| is connected
+            to node |j|. If no connectivity matrix is given, every node is
+            connected to every node **(including itself)**.
+        node_labels (tuple[str]): Human readable labels for each node in the
+            network.
+
     Example:
         In a 3-node network, ``a_network.tpm[(0, 0, 1)]`` gives the transition
         probabilities for each node at |t_0| given that state at |t_{-1}| was
         |N_0 = 0, N_1 = 0, N_2 = 1|.
-
-    Args:
-        tpm (np.ndarray): See the corresponding attribute.
-
-    Keyword Args:
-        connectivity_matrix (array or sequence): A square binary adjacency
-            matrix indicating the connections between nodes in the network.
-            ``connectivity_matrix[i][j] == 1`` means that node |i| is connected
-            to node |j|. If no connectivity matrix is given, every node is
-            connected to every node **(including itself)**.
-        node_labels (tuple(str)): Human readable labels for each node in the
-            network.
-
-    Attributes:
-        tpm (np.ndarray):
-            The network's transition probability matrix. It can be provided in
-            either state-by-node (either 2-D or N-D) or state-by-state form. In
-            either form, row indices must follow the **LOLI** convention (see
-            discussion in the |examples| module), and in state-by-state form,
-            so must column indices. If given in state-by-node form, it can be
-            either 2-dimensional, so that ``tpm[i]`` gives the probabilities of
-            each node being on if the past state is encoded by |i| according to
-            **LOLI**, or in N-D form, so that ``tpm[(0, 0, 1)]`` gives the
-            probabilities of each node being on if the past state is |N_0 = 0,
-            N_1 = 0, N_2 = 1|. The shape of the 2-dimensional form of a
-            state-by-node TPM must be ``(S, N)``, and the shape of the N-D form
-            of the TPM must be ``[2] * N + [N]``, where ``S`` is the number of
-            states and ``N`` is the number of nodes in the network.
-        cm (np.ndarray):
-            A square binary adjacency matrix indicating the connections between
-            nodes in the network.
-        connectivity_matrix (np.ndarray):
-            Alias for `Network.cm`.
-        size (int):
-            The number of nodes in the network.
-        num_states (int):
-            The number of possible states of the network.
     """
 
     # TODO make tpm also optional when implementing logical network definition
     def __init__(self, tpm, connectivity_matrix=None, node_labels=None,
                  perturb_vector=None, purview_cache=None):
-        self.tpm = tpm
+
+        self._tpm, self._tpm_hash = self._build_tpm(tpm)
+        self._cm, self._cm_hash = self._build_cm(connectivity_matrix)
         self._node_indices = tuple(range(self.size))
         self._node_labels = node_labels
-        self.cm = connectivity_matrix
         self.perturb_vector = perturb_vector
         self.purview_cache = purview_cache or cache.PurviewCache()
 
         validate.network(self)
 
     @property
+    def tpm(self):
+        """np.ndarray: The network's transition probability matrix, in |N-D|
+        form."""
+        return self._tpm
+
+    def _build_tpm(self, tpm):
+        """Validate the TPM passed by the user and convert to |N-D| form. """
+        tpm = np.array(tpm)
+
+        validate.tpm(tpm)
+
+        # Convert to N-D state-by-node form
+        if utils.state_by_state(tpm):
+            tpm = convert.state_by_state2state_by_node(tpm)
+        else:
+            tpm = convert.to_n_dimensional(tpm)
+
+        # Make the underlying attribute immutable.
+        tpm.flags.writeable = False
+
+        return (tpm, utils.np_hash(tpm))
+
+    @property
+    def cm(self):
+        """np.ndarray: The network's connectivity matrix.
+
+        A square binary adjacency matrix indicating the connections between
+        nodes in the network.
+        """
+        return self._cm
+
+    def _build_cm(self, cm):
+        """Convert the passed CM to the proper format, or construct the
+        unitary CM if none was provided."""
+        if cm is None:
+            # Assume all are connected.
+            cm = np.ones((self.size, self.size))
+        else:
+            cm = np.array(cm)
+
+        # Make the underlying attribute immutable.
+        cm.flags.writeable = False
+
+        return (cm, utils.np_hash(cm))
+
+    @property
+    def connectivity_matrix(self):
+        """np.ndarray: Alias for `Network.cm`."""
+        return self._cm
+
+    @property
     def size(self):
+        """int: The number of nodes in the network."""
         return self.tpm.shape[-1]
 
     # TODO extend to nonbinary nodes
     @property
     def num_states(self):
+        """int: The number of possible states of the network."""
         return 2 ** self.size
 
     @property
     def node_indices(self):
+        """tuple[int]: The indices of nodes in the network.
+
+        This is ``0..network.size``.
+        """
         return self._node_indices
 
     @property
     def node_labels(self):
+        """tuple[str]: The labels of nodes in the network."""
         if self._node_labels is None:
             self._node_labels = tuple("n{}".format(i)
                                       for i in self.node_indices)
         return self._node_labels
-
-    @property
-    def tpm(self):
-        return self._tpm
-
-    @tpm.setter
-    def tpm(self, tpm):
-        # Cast TPM to np.array.
-        tpm = np.array(tpm)
-        # Validate TPM.
-        # The TPM can be either 2-dimensional or in N-D form, where transition
-        # probabilities can be indexed by state-tuples.
-        validate.tpm(tpm)
-        # Convert to N-D state-by-node if we were given a square state-by-state
-        # TPM. Otherwise, force conversion to N-D format.
-        if utils.state_by_state(tpm):
-            self._tpm = convert.state_by_state2state_by_node(tpm)
-        else:
-            self._tpm = convert.to_n_dimensional(tpm)
-        # Make the underlying attribute immutable.
-        self._tpm.flags.writeable = False
-        # Update hash.
-        self._tpm_hash = utils.np_hash(self.tpm)
-
-    @property
-    def cm(self):
-        """The network's connectivity matrix."""
-        return self._cm
-
-    @cm.setter
-    def cm(self, cm):
-        if cm is not None:
-            self._cm = np.array(cm)
-        else:
-            # If none was provided, assume all are connected.
-            self._cm = np.ones((self.size, self.size))
-        # Make the underlying attribute immutable.
-        self._cm.flags.writeable = False
-        # Update hash.
-        self._cm_hash = utils.np_hash(self.cm)
-
-    @property
-    def connectivity_matrix(self):
-        """Alias for `connectivity_matrix`."""
-        return self.cm
-
-    @connectivity_matrix.setter
-    def connectivity_matrix(self, cm):
-        self.cm = cm
 
     @property
     def perturb_vector(self):
@@ -192,11 +193,11 @@ class Network:
 
         Args:
             direction (str): |past| or |future|
-            mechanism (tuple(int)): The mechanism which all purviews are
+            mechanism (tuple[int]): The mechanism which all purviews are
                 checked for reducibility over.
 
         Returns:
-            list(tuple(int)): All purviews which are irreducible over
+            list[tuple[int]]: All purviews which are irreducible over
                 ``mechanism``.
         """
         all_purviews = utils.powerset(self._node_indices)
@@ -234,7 +235,8 @@ class Network:
         return {
             'tpm': self.tpm,
             'cm': self.cm,
-            'size': self.size
+            'size': self.size,
+            'labels': self.node_labels
         }
 
 
@@ -244,11 +246,11 @@ def irreducible_purviews(cm, direction, mechanism, purviews):
     Args:
         cm (np.ndarray): A |N x N| connectivity matrix.
         direction (str): |past| or |future|.
-        purviews (list(tuple(int))): The purviews to check.
-        mechanism (tuple(int)): The mechanism in question.
+        purviews (list[tuple[int]]): The purviews to check.
+        mechanism (tuple[int]): The mechanism in question.
 
     Returns:
-        list(tuple(int)): All purviews in ``purviews`` which are not reducible
+        list[tuple[int]]: All purviews in ``purviews`` which are not reducible
             over ``mechanism``.
     """
     def reducible(purview):
@@ -269,11 +271,13 @@ def from_json(filename):
         filename (str): A path to a JSON file representing a network.
 
     Returns:
-       network (``Network``): The corresponding PyPhi network object.
+       |Network|: The corresponding PyPhi network object.
     """
     with open(filename) as f:
-        network_dictionary = json.load(f)
-    tpm = network_dictionary['tpm']
-    cm = network_dictionary['cm']
-    network = Network(tpm, connectivity_matrix=cm)
-    return network
+        loaded = json.load(f)
+
+    tpm = loaded['tpm']
+    cm = loaded['cm']
+    labels = loaded['labels']
+
+    return Network(tpm, connectivity_matrix=cm, node_labels=labels)